requirements.txt
python-pptx
langchain-openai
Pillow
pexels-api
flask
python-dotenv
langchain
openai
llama-index
pypdf
chromadb
tiktoken
streamlit
nltk
pydantic
huggingface_hub
httpx
coloredlogs
numpy
pandas
pytz
PyYAML
requests
rich
sympy
tqdm
typing-extensions
gunicorn